[refact] rename ssd constants
