+++ ssd/src/main/java/com/samsung/buffer/handler/ReadCommandHandler.java
@@ -5,7 +5,7 @@ import com.samsung.common.CmdData;
 
 import java.util.Map;
 
-import static com.samsung.common.SSDConstant.EMPTY_VALUE;
+import static com.samsung.common.SSDConstants.EMPTY_VALUE;
 
 public class ReadCommandHandler implements CommandHandler {
     private final Map<Integer, String> memory;
