+++ ssd/src/main/java/com/samsung/common/SSDConstants.java
@@ -1,13 +1,12 @@
 package com.samsung.common;
 
-public class SSDConstant {
-    private SSDConstant() {}
+public class SSDConstants {
+    private SSDConstants() {}
 
     public static final String ERROR_MESSAGE = "ERROR";
     public static final String EMPTY_VALUE = "0x00000000";
     public static final int MAX_LBA = 99;
     public static final int MAX_ERASE_SIZE = 10;
-
     public static final int MIN_BUFFER_INDEX = 1;
     public static final int MAX_BUFFER_INDEX = 5;
 }
